+++ packages/python/tests/ci/config/test_free_flow.py
@@ -7,7 +7,6 @@ from m.ci.config import (
     MFlowConfig,
     Workflow,
     read_config,
-    read_workflow,
 )
 from m.core import issue
 from m.core.fp import Good
@@ -31,15 +30,6 @@ class ConfigFreeFlowTest(FpTestCase):
         m_flow=MFlowConfig(),
     )
 
-    def test_workflow_config(self):
-        result = read_workflow('free-flow')
-        workflow = self.assert_ok(result)
-        self.assertEqual(workflow, Workflow.free_flow)
-
-    def test_workflow_config_fail(self):
-        result = read_workflow('unknown-flow')
-        self.assert_issue(result, 'invalid workflow')
-
     @patch('m.core.json.read_json')
     def test_read_config_fail(self, read_json_mock):
         read_json_mock.return_value = issue('made up issue')
@@ -52,15 +42,12 @@ class ConfigFreeFlowTest(FpTestCase):
             result = read_config('m')
             self.assert_issue(result, 'read_config failure')
             err = cast(Issue, cast(Issue, result.value).cause)
-            self.assertEqual(err.message, 'multi_get key retrieval failure')
-            if isinstance(err.context, list):
-                msgs = {x['cause']['message'] for x in err.context}
-                self.assertSetEqual(
-                    msgs, set([
-                        '`owner` path was not found',
-                        '`repo` path was not found',
-                    ]),
-                )
+            self.assertEqual(err.message, 'pydantic validation error')
+            if err.cause:
+                msgs = str(err.cause)
+                assert '2 validation errors for Config' in msgs
+                assert 'owner\n  field required' in msgs
+                assert 'repo\n  field required' in msgs
             else:
                 raise AssertionError('issue data should be a string')